+++ src/App.jsx
@@ -1,5 +1,5 @@
 import './App.scss';
-import { PostList } from './components/PostList/PostList';
+import { PostList } from './components/PostList';
 
 import postsFromServer from './api/posts.json';
 import usersFromServer from './api/users.json';
