+++ src/components/PostList/PostList.jsx
@@ -1,4 +1,4 @@
-import { PostInfo } from '../PostInfo/PostInfo';
+import { PostInfo } from '../PostInfo';
 
 export const PostList = ({ posts }) => (
   <div className="PostList">
